--- a/staging/src/k8s.io/apiserver/pkg/storage/cacher/watch_cache.go
+++ b/staging/src/k8s.io/apiserver/pkg/storage/cacher/watch_cache.go
@@ -25,6 +25,7 @@ import (
 	"time"
 
 	"k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/fields"
 	"k8s.io/apimachinery/pkg/labels"
 	"k8s.io/apimachinery/pkg/runtime"
@@ -151,6 +152,8 @@ type watchCache struct {
 	// Requests progress notification if there are requests waiting for watch
 	// to be fresh
 	waitingUntilFresh *conditionalProgressRequester
+
+	storeSnapshots *storeSnapshotter
 }
 
 func newWatchCache(
@@ -181,7 +184,9 @@ func newWatchCache(
 		versioner:           versioner,
 		groupResource:       groupResource,
 		waitingUntilFresh:   progressRequester,
+		storeSnapshots:      newStoreSnapshotter(),
 	}
+
 	metrics.WatchCacheCapacity.WithLabelValues(groupResource.String()).Set(float64(wc.capacity))
 	wc.cond = sync.NewCond(wc.RLocker())
 	wc.indexValidator = wc.isIndexValidLocked
@@ -306,6 +311,8 @@ func (w *watchCache) processEvent(event watch.Event, resourceVersion uint64, upd
 func (w *watchCache) updateCache(event *watchCacheEvent) {
 	w.resizeCacheLocked(event.RecordTime)
 	if w.isCacheFullLocked() {
+		oldestRV := w.cache[w.startIndex%w.capacity].ResourceVersion
+		w.storeSnapshots.Clean(oldestRV)
 		// Cache is full - remove the oldest element.
 		w.startIndex++
 		w.removedEventSinceRelist = true
@@ -451,7 +458,7 @@ func (s sortableStoreElements) Swap(i, j int) {
 
 // WaitUntilFreshAndList returns list of pointers to `storeElement` objects along
 // with their ResourceVersion and the name of the index, if any, that was used.
-func (w *watchCache) WaitUntilFreshAndList(ctx context.Context, resourceVersion uint64, key string, matchValues []storage.MatchValue) (resp listResp, index string, err error) {
+func (w *watchCache) WaitUntilFreshAndList(ctx context.Context, resourceVersion uint64, key string, opts storage.ListOptions, matchValues []storage.MatchValue) (resp listResp, index string, err error) {
 	requestWatchProgressSupported := etcdfeature.DefaultFeatureSupportChecker.Supports(storage.RequestWatchProgress)
 	if utilfeature.DefaultFeatureGate.Enabled(features.ConsistentListFromCache) && requestWatchProgressSupported && w.notFresh(resourceVersion) {
 		w.waitingUntilFresh.Add()
@@ -465,38 +472,133 @@ func (w *watchCache) WaitUntilFreshAndList(ctx context.Context, resourceVersion 
 	if err != nil {
 		return listResp{}, "", err
 	}
-	// This isn't the place where we do "final filtering" - only some "prefiltering" is happening here. So the only
-	// requirement here is to NOT miss anything that should be returned. We can return as many non-matching items as we
-	// want - they will be filtered out later. The fact that we return less things is only further performance improvement.
-	// TODO: if multiple indexes match, return the one with the fewest items, so as to do as much filtering as possible.
-	for _, matchValue := range matchValues {
-		if result, err := w.store.ByIndex(matchValue.IndexName, matchValue.Value); err == nil {
-			result, err = filterPrefixAndOrder(key, result)
+	nonEmptyResourceVersion := len(opts.ResourceVersion) > 0 && opts.ResourceVersion != "0"
+	isLegacyResourceVersionMatchExact := opts.ResourceVersionMatch == "" && opts.Predicate.Limit > 0 && nonEmptyResourceVersion
+	switch {
+	case opts.ResourceVersionMatch == metav1.ResourceVersionMatchExact || isLegacyResourceVersionMatchExact:
+		store, ok := w.storeSnapshots.Get(resourceVersion)
+		if !ok {
+			err = errors.NewResourceExpired(fmt.Sprintf("too old resource version: %d", resourceVersion))
+			return listResp{}, "", err
+		}
+		return paginatedRead(store, key, "", resourceVersion, opts), "", nil
+	case opts.ResourceVersionMatch != "" && opts.ResourceVersionMatch != metav1.ResourceVersionMatchNotOlderThan:
+		return listResp{}, "", fmt.Errorf("unsupported")
+	case len(opts.Predicate.Continue) > 0:
+		continueKey, continueRV, err := storage.DecodeContinue(opts.Predicate.Continue, key)
+		if err != nil {
+			return listResp{}, "", errors.NewBadRequest(fmt.Sprintf("invalid continue token: %v", err))
+		}
+		if nonEmptyResourceVersion {
+			return listResp{}, "", errors.NewBadRequest("specifying resource version is not allowed when using continue")
+		}
+		store, ok := w.storeSnapshots.Get(uint64(continueRV))
+		if !ok {
+			err = errors.NewResourceExpired(fmt.Sprintf("too old resource version: %d", continueRV))
+			return listResp{}, "", err
+		}
+		return paginatedRead(store, key, continueKey, uint64(continueRV), opts), "", nil
+	default:
+		orderedStore, ordered := w.store.(orderedLister)
+		if opts.Predicate.Limit > 0 && ordered {
+			w.storeSnapshots.Set(w.resourceVersion, orderedStore)
+		}
+
+		// This isn't the place where we do "final filtering" - only some "prefiltering" is happening here. So the only
+		// requirement here is to NOT miss anything that should be returned. We can return as many non-matching items as we
+		// want - they will be filtered out later. The fact that we return less things is only further performance improvement.
+		// TODO: if multiple indexes match, return the one with the fewest items, so as to do as much filtering as possible.
+		for _, matchValue := range matchValues {
+			if result, err := w.store.ByIndex(matchValue.IndexName, matchValue.Value); err == nil {
+				result, err = filterPrefixAndOrder(key, result)
+				return listResp{
+					Items:           result,
+					HasMore:         false,
+					ItemCount:       int64(len(result)),
+					ResourceVersion: w.resourceVersion,
+				}, matchValue.IndexName, err
+			}
+		}
+		if ordered {
+			result, hasMore := orderedStore.ListPrefix(key, "", 0)
 			return listResp{
 				Items:           result,
-				HasMore:         false,
+				HasMore:         hasMore,
 				ItemCount:       int64(len(result)),
 				ResourceVersion: w.resourceVersion,
-			}, matchValue.IndexName, err
+			}, "", nil
 		}
-	}
-	if store, ok := w.store.(orderedLister); ok {
-		result, hasMore := store.ListPrefix(key, "", 0)
+		result := w.store.List()
+		result, err = filterPrefixAndOrder(key, result)
 		return listResp{
 			Items:           result,
-			HasMore:         hasMore,
+			HasMore:         false,
 			ItemCount:       int64(len(result)),
 			ResourceVersion: w.resourceVersion,
-		}, "", nil
+		}, "", err
 	}
-	result := w.store.List()
-	result, err = filterPrefixAndOrder(key, result)
-	return listResp{
-		Items:           result,
-		HasMore:         false,
-		ItemCount:       int64(len(result)),
-		ResourceVersion: w.resourceVersion,
-	}, "", err
+}
+
+func paginatedRead(store orderedLister, key, continueKey string, resourceVersion uint64, opts storage.ListOptions) listResp {
+	listLimit := int(computeListLimit(opts))
+	readLimit := listLimit
+	allItems := make([]interface{}, 0, listLimit)
+	var moreInFilter bool
+	for {
+		items, moreInStore := store.ListPrefix(key, continueKey, readLimit)
+		if opts.Predicate.Empty() {
+			// Need to provide accurate ItemCount for empty predicate
+			itemCount := len(items)
+			if listLimit > 0 {
+				itemCount = store.Count(key, continueKey)
+			}
+			return listResp{
+				Items:           items,
+				HasMore:         moreInStore,
+				ResourceVersion: resourceVersion,
+				ItemCount:       int64(itemCount),
+			}
+		}
+		if listLimit <= 0 {
+			return listResp{
+				Items:           items,
+				ResourceVersion: resourceVersion,
+				HasMore:         moreInStore,
+			}
+		}
+		items, moreInFilter = filterUpToLimit(items, listLimit-len(allItems), opts.Predicate)
+		// TODO: Get continuation token before filtering.
+		if len(items) > 0 {
+			continueKey = items[len(items)-1].(*storeElement).Key + "\x00"
+		}
+		allItems = append(allItems, items...)
+		if !moreInStore || len(allItems) >= listLimit {
+			return listResp{
+				Items:           allItems,
+				ResourceVersion: resourceVersion,
+				HasMore:         moreInStore || moreInFilter,
+			}
+		}
+		readLimit *= 2
+	}
+}
+
+func filterUpToLimit(items []interface{}, limit int, pred storage.SelectionPredicate) (matchedItems []interface{}, hasMore bool) {
+	matchedItems = make([]interface{}, 0, limit)
+	for _, obj := range items {
+		elem := obj.(*storeElement)
+		if !pred.MatchesObjectAttributes(elem.Labels, elem.Fields) {
+			continue
+		}
+		// TODO: Might be worth to lookup one more item to provide more accurate HasMore.
+		if len(matchedItems) < limit {
+			matchedItems = append(matchedItems, obj)
+		} else {
+			hasMore = true
+			return matchedItems, hasMore
+		}
+	}
+	return matchedItems, hasMore
 }
 
 func filterPrefixAndOrder(prefix string, items []interface{}) ([]interface{}, error) {